feat: refactor PrinterSelector component; replace selectedPrinterId prop with file prop for improved printer selection handling
