+++ src/components/Model/PrinterSelectorModel.tsx
@@ -1,5 +1,3 @@
-import { useEffect, useState } from 'react';
-import { PrinterInfo } from '../PrinterCard';
 import { Box, Button } from '@mui/material';
 import {
   CommandDialog,
@@ -8,73 +6,36 @@ import {
   CommandItem,
   CommandList,
 } from '../MUI/Command';
-
+import { PrinterInfo } from '../PrinterCard';
 import { IPrinter } from '../../types/printer';
-import { useDispatch, useSelector } from 'react-redux';
-import { UpdateValueById } from '../../store/customizeFilesDetails/CustomizationSlice';
-import { RootState } from '../../store/types';
+import { FileDataDB } from '../../types/uploadFiles';
+import { usePrinterSelector } from '../../hook/usePrinterSelector';
 
 export default function PrinterSelector({
   printersData,
-  selectedPrinterId,
+  file,
 }: {
   printersData: IPrinter[];
-  selectedPrinterId?: string | null;
+  file: FileDataDB | undefined;
 }) {
-  const [open, setOpen] = useState(false);
-  const [selectedPrinter, setSelectedPrinter] = useState<IPrinter | null>(null);
-  const [searchValue, setSearchValue] = useState('');
-  // Redux
-  const { activeFileId } = useSelector(
-    (state: RootState) => state.customization
-  );
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    const down = (e: KeyboardEvent) => {
-      if (e.key === 'k' && (e.metaKey || e.ctrlKey)) {
-        e.preventDefault();
-        setOpen((open) => !open);
-      }
-    };
-    document.addEventListener('keydown', down);
-    return () => document.removeEventListener('keydown', down);
-  }, []);
-
-  const handleSelect = (printer: IPrinter) => {
-    setSelectedPrinter(printer);
-    dispatch(
-      UpdateValueById({
-        id: activeFileId as string,
-        data: { printerId: printer._id },
-      })
-    );
-    setOpen(false);
-  };
-
-  const filteredPrinters = printersData.filter((p) =>
-    p.name.toLowerCase().includes(searchValue.toLowerCase())
-  );
-
-  useEffect(() => {
-    if (selectedPrinterId && printersData.length > 0) {
-      const printer = printersData.find((p) => p._id === selectedPrinterId);
-      if (printer) {
-        setSelectedPrinter(printer);
-      }
-    } else {
-      setSelectedPrinter(null);
-    }
-  }, [selectedPrinterId, printersData]);
+  
+  const {
+    open,
+    setOpen,
+    selectedPrinter,
+    handleSelect,
+    searchValue,
+    setSearchValue,
+    filteredPrinters,
+    isDisabled,
+  } = usePrinterSelector(printersData, file);
 
   return (
-    <Box
-      sx={{ display: 'flex', flexDirection: 'column', alignItems: 'center' }}
-    >
+    <Box sx={{ display: 'flex', flexDirection: 'column', alignItems: 'center' }}>
       <Button
         variant="outlined"
         onClick={() => setOpen(true)}
-        // endIcon={<UnfoldMoreIcon />}
+        disabled={isDisabled}
         sx={{
           width: '100%',
           justifyContent: 'space-between',
